Add judge and registrars to reception and assingment step
