+++ apps/judicial-system/web/messages/RestrictionCases/Court/receptionAndAssignment.ts
@@ -1,3 +1,4 @@
+import { defineMessages } from '@formatjs/intl'
 import { defineMessage } from 'react-intl'
 
 export const rcReceptionAndAssignment = {
@@ -5,6 +6,42 @@ export const rcReceptionAndAssignment = {
     id: 'judicial.system.restriction_cases:reception_and_assignment.title',
     defaultMessage: 'Móttaka og úthlutun',
     description:
-      'Notaður sem titill á Móttaka og úthlutun skjá í gæsluvarðhalds- og farbannsmálum.',
+      'Notaður sem titill á Móttaka og úthlutun skrefi í gæsluvarðhalds- og farbannsmálum.',
   }),
+  sections: {
+    setJudge: defineMessages({
+      title: {
+        id:
+          'judicial.system.restriction_cases:reception_and_assignment.set_judge.title',
+        defaultMessage: 'Dómari',
+        description:
+          'Notaður sem titill fyrir "dómari" hlutann á Móttaka og úthlutun skrefi í gæsluvarðhalds- og farbannsmálum.',
+      },
+      tooltip: {
+        id:
+          'judicial.system.restriction_cases:reception_and_assignment.set_judge.tooltip',
+        defaultMessage:
+          'Dómarinn sem er valinn hér verður skráður á málið og mun fá tilkynningar sendar í tölvupóst. Eingöngu skráður dómari getur svo undirritað úrskurð.',
+        description:
+          'Notaður sem upplýsingatexti í upplýsingasvæði við "dómari" titlinn á Móttaka og úthlutun skrefi í gæsluvarðhalds- og farbannsmálum.',
+      },
+    }),
+    setRegistrar: defineMessages({
+      title: {
+        id:
+          'judicial.system.restriction_cases:reception_and_assignment.set_registrar.title',
+        defaultMessage: 'Dómritari',
+        description:
+          'Notaður sem titill fyrir "Dómritari" hlutann á Móttaka og úthlutun skrefi í gæsluvarðhalds- og farbannsmálum.',
+      },
+      tooltip: {
+        id:
+          'judicial.system.restriction_cases:reception_and_assignment.set_registrar.tooltip',
+        defaultMessage:
+          'Dómritari sem er valinn hér verður skráður á málið og mun fá tilkynningar sendar í tölvupósti.',
+        description:
+          'Notaður sem upplýsingatexti í upplýsingasvæði við "dómritari" titlinn á Móttaka og úthlutun skrefi í gæsluvarðhalds- og farbannsmálum.',
+      },
+    }),
+  },
 }
