+++ apps/judicial-system/web/src/routes/Court/RestrictionRequest/HearingArrangements/HearingArrangements.tsx
@@ -1,18 +1,8 @@
 import React, { useContext, useEffect, useState } from 'react'
 import { useIntl } from 'react-intl'
-import { useQuery } from '@apollo/client'
 import { useRouter } from 'next/router'
-import { ValueType } from 'react-select/src/types'
 
-import {
-  Box,
-  Input,
-  Select,
-  Text,
-  Option,
-  Tooltip,
-  AlertMessage,
-} from '@island.is/island-ui/core'
+import { Box, Input, Text, AlertMessage } from '@island.is/island-ui/core'
 import {
   FormFooter,
   PageLayout,
@@ -22,25 +12,16 @@ import {
   Modal,
 } from '@island.is/judicial-system-web/src/components'
 import { isCourtHearingArrangemenstStepValidRC } from '@island.is/judicial-system-web/src/utils/validate'
-import {
-  CaseType,
-  NotificationType,
-  UserRole,
-} from '@island.is/judicial-system/types'
-import type { User } from '@island.is/judicial-system/types'
+import { CaseType, NotificationType } from '@island.is/judicial-system/types'
 import {
   JudgeSubsections,
-  ReactSelectOption,
   Sections,
-  UserData,
 } from '@island.is/judicial-system-web/src/types'
 import {
   validateAndSendToServer,
   removeTabsValidateAndSet,
-  setAndSendToServer,
   setAndSendDateToServer,
 } from '@island.is/judicial-system-web/src/utils/formHelper'
-import { UsersQuery } from '@island.is/judicial-system-web/src/utils/mutations'
 import { DateTime } from '@island.is/judicial-system-web/src/components'
 import { useCase } from '@island.is/judicial-system-web/src/utils/hooks'
 import { FormContext } from '@island.is/judicial-system-web/src/components/FormProvider/FormProvider'
@@ -72,42 +53,6 @@ export const HearingArrangements: React.FC = () => {
   } = useCase()
   const { formatMessage } = useIntl()
 
-  const { data: userData, loading: userLoading } = useQuery<UserData>(
-    UsersQuery,
-    {
-      fetchPolicy: 'no-cache',
-      errorPolicy: 'all',
-    },
-  )
-
-  const judges = (userData?.users ?? [])
-    .filter(
-      (user: User) =>
-        user.role === UserRole.JUDGE &&
-        user.institution?.id === workingCase?.court?.id,
-    )
-    .map((judge: User) => {
-      return { label: judge.name, value: judge.id }
-    })
-
-  const registrars = (userData?.users ?? [])
-    .filter(
-      (user: User) =>
-        user.role === UserRole.REGISTRAR &&
-        user.institution?.id === workingCase?.court?.id,
-    )
-    .map((registrar: User) => {
-      return { label: registrar.name, value: registrar.id }
-    })
-
-  const defaultJudge = judges?.find(
-    (judge: Option) => judge.value === workingCase?.judge?.id,
-  )
-
-  const defaultRegistrar = registrars?.find(
-    (registrar: Option) => registrar.value === workingCase?.registrar?.id,
-  )
-
   useEffect(() => {
     document.title = 'Fyrirtaka - Réttarvörslugátt'
   }, [])
@@ -123,32 +68,6 @@ export const HearingArrangements: React.FC = () => {
       setWorkingCase(theCase)
     }
   }, [autofill, isCaseUpToDate, setWorkingCase, workingCase])
-
-  const setJudge = (id: string) => {
-    if (workingCase) {
-      setAndSendToServer('judgeId', id, workingCase, setWorkingCase, updateCase)
-
-      const judge = userData?.users.find((j) => j.id === id)
-
-      setWorkingCase({ ...workingCase, judge: judge })
-    }
-  }
-
-  const setRegistrar = (id?: string) => {
-    if (workingCase) {
-      setAndSendToServer(
-        'registrarId',
-        id,
-        workingCase,
-        setWorkingCase,
-        updateCase,
-      )
-
-      const registrar = userData?.users.find((r) => r.id === id)
-
-      setWorkingCase({ ...workingCase, registrar })
-    }
-  }
 
   const handleNextButtonClick = () => {
     if (
@@ -169,7 +88,7 @@ export const HearingArrangements: React.FC = () => {
         workingCase?.parentCase ? Sections.JUDGE_EXTENSION : Sections.JUDGE
       }
       activeSubSection={JudgeSubsections.HEARING_ARRANGEMENTS}
-      isLoading={isLoadingWorkingCase || userLoading}
+      isLoading={isLoadingWorkingCase}
       notFound={caseNotFound}
     >
       <FormContentContainer>
@@ -189,50 +108,6 @@ export const HearingArrangements: React.FC = () => {
         </Box>
         <Box component="section" marginBottom={7}>
           <CaseInfo workingCase={workingCase} userRole={user?.role} />
-        </Box>
-        <Box component="section" marginBottom={5}>
-          <Box marginBottom={3}>
-            <Text as="h3" variant="h3">
-              {`${formatMessage(m.sections.setJudge.title)} `}
-              <Tooltip text={formatMessage(m.sections.setJudge.tooltip)} />
-            </Text>
-          </Box>
-          <Select
-            name="judge"
-            label="Veldu dómara"
-            placeholder="Velja héraðsdómara"
-            value={defaultJudge}
-            options={judges}
-            onChange={(selectedOption: ValueType<ReactSelectOption>) =>
-              setJudge((selectedOption as ReactSelectOption).value.toString())
-            }
-            required
-          />
-        </Box>
-        <Box component="section" marginBottom={5}>
-          <Box marginBottom={3}>
-            <Text as="h3" variant="h3">
-              {`${formatMessage(m.sections.setRegistrar.title)} `}
-              <Tooltip text={formatMessage(m.sections.setRegistrar.tooltip)} />
-            </Text>
-          </Box>
-          <Select
-            name="registrar"
-            label="Veldu dómritara"
-            placeholder="Velja dómritara"
-            value={defaultRegistrar}
-            options={registrars}
-            onChange={(selectedOption: ValueType<ReactSelectOption>) => {
-              if (selectedOption) {
-                setRegistrar(
-                  (selectedOption as ReactSelectOption).value.toString(),
-                )
-              } else {
-                setRegistrar(undefined)
-              }
-            }}
-            isClearable
-          />
         </Box>
         <Box component="section" marginBottom={8}>
           <Box marginBottom={2}>
