+++ apps/judicial-system/web/src/routes/Court/RestrictionRequest/ReceptionAndAssignment/ReceptionAndAssignmentForm.tsx
@@ -1,12 +1,16 @@
 import React from 'react'
 import { useIntl } from 'react-intl'
+import { ValueType } from 'react-select/src/types'
 
-import { Box, Text } from '@island.is/island-ui/core'
+import { Box, Select, Text, Tooltip, Option } from '@island.is/island-ui/core'
 import { FormContentContainer } from '@island.is/judicial-system-web/src/components'
-import { Case } from '@island.is/judicial-system/types'
+import { Case, User, UserRole } from '@island.is/judicial-system/types'
+import { ReactSelectOption } from '@island.is/judicial-system-web/src/types'
 import { rcReceptionAndAssignment as m } from '@island.is/judicial-system-web/messages/RestrictionCases/Court/receptionAndAssignment'
 
 import CourtCaseNumber from '../../SharedComponents/CourtCaseNumber/CourtCaseNumber'
+import { setAndSendToServer } from '@island.is/judicial-system-web/src/utils/formHelper'
+import { useCase } from '@island.is/judicial-system-web/src/utils/hooks'
 
 interface Props {
   workingCase: Case
@@ -18,6 +22,7 @@ interface Props {
   setCourtCaseNumberEM: React.Dispatch<React.SetStateAction<string>>
   isCreatingCourtCase: boolean
   receiveCase: (wc: Case, courtCaseNumber: string) => void
+  users?: User[]
 }
 
 const ReceptionAndAssignementForm: React.FC<Props> = (props) => {
@@ -31,8 +36,64 @@ const ReceptionAndAssignementForm: React.FC<Props> = (props) => {
     handleCreateCourtCase,
     isCreatingCourtCase,
     receiveCase,
+    users,
   } = props
   const { formatMessage } = useIntl()
+  const { updateCase } = useCase()
+
+  const judges = (users ?? [])
+    .filter(
+      (user: User) =>
+        user.role === UserRole.JUDGE &&
+        user.institution?.id === workingCase?.court?.id,
+    )
+    .map((judge: User) => {
+      return { label: judge.name, value: judge.id }
+    })
+
+  const registrars = (users ?? [])
+    .filter(
+      (user: User) =>
+        user.role === UserRole.REGISTRAR &&
+        user.institution?.id === workingCase?.court?.id,
+    )
+    .map((registrar: User) => {
+      return { label: registrar.name, value: registrar.id }
+    })
+
+  const defaultJudge = judges?.find(
+    (judge: Option) => judge.value === workingCase?.judge?.id,
+  )
+
+  const defaultRegistrar = registrars?.find(
+    (registrar: Option) => registrar.value === workingCase?.registrar?.id,
+  )
+
+  const setJudge = (id: string) => {
+    if (workingCase) {
+      setAndSendToServer('judgeId', id, workingCase, setWorkingCase, updateCase)
+
+      const judge = users?.find((j) => j.id === id)
+
+      setWorkingCase({ ...workingCase, judge: judge })
+    }
+  }
+
+  const setRegistrar = (id?: string) => {
+    if (workingCase) {
+      setAndSendToServer(
+        'registrarId',
+        id,
+        workingCase,
+        setWorkingCase,
+        updateCase,
+      )
+
+      const registrar = users?.find((r) => r.id === id)
+
+      setWorkingCase({ ...workingCase, registrar })
+    }
+  }
 
   return (
     <FormContentContainer>
@@ -52,6 +113,50 @@ const ReceptionAndAssignementForm: React.FC<Props> = (props) => {
           handleCreateCourtCase={handleCreateCourtCase}
           isCreatingCourtCase={isCreatingCourtCase}
           receiveCase={receiveCase}
+        />
+      </Box>
+      <Box component="section" marginBottom={5}>
+        <Box marginBottom={3}>
+          <Text as="h3" variant="h3">
+            {`${formatMessage(m.sections.setJudge.title)} `}
+            <Tooltip text={formatMessage(m.sections.setJudge.tooltip)} />
+          </Text>
+        </Box>
+        <Select
+          name="judge"
+          label="Veldu dómara"
+          placeholder="Velja héraðsdómara"
+          value={defaultJudge}
+          options={judges}
+          onChange={(selectedOption: ValueType<ReactSelectOption>) =>
+            setJudge((selectedOption as ReactSelectOption).value.toString())
+          }
+          required
+        />
+      </Box>
+      <Box component="section" marginBottom={10}>
+        <Box marginBottom={3}>
+          <Text as="h3" variant="h3">
+            {`${formatMessage(m.sections.setRegistrar.title)} `}
+            <Tooltip text={formatMessage(m.sections.setRegistrar.tooltip)} />
+          </Text>
+        </Box>
+        <Select
+          name="registrar"
+          label="Veldu dómritara"
+          placeholder="Velja dómritara"
+          value={defaultRegistrar}
+          options={registrars}
+          onChange={(selectedOption: ValueType<ReactSelectOption>) => {
+            if (selectedOption) {
+              setRegistrar(
+                (selectedOption as ReactSelectOption).value.toString(),
+              )
+            } else {
+              setRegistrar(undefined)
+            }
+          }}
+          isClearable
         />
       </Box>
     </FormContentContainer>
